relocated trait for relations
