+++ src/Sulu/Component/Rest/RestHelper.php
@@ -10,6 +10,7 @@
 
 namespace Sulu\Component\Rest;
 
+use Sulu\Component\Persistence\RelationTrait;
 use Sulu\Component\Rest\ListBuilder\ListBuilderInterface;
 use Sulu\Component\Rest\ListBuilder\ListRestHelper;
 use Sulu\Component\Rest\ListBuilder\ListRestHelperInterface;
